fix: extract select options from campaigns index
